Implement CSV upload functionality for glucose data processing and add Celery task support

+++ src/modules/api/models/glucose.py
@@ -25,7 +25,7 @@ class Glucose(TimeStampedModel):
     user_id = models.CharField("User ID", max_length=40, blank=False, null=False)
     record_date = models.DateField("Date")
     record_time = models.TimeField("Time")
-    glucose_value = models.FloatField("Glucose Value mg/dL", blank=False, null=False)
+    glucose_value = models.IntegerField("Glucose Value mg/dL", blank=False, null=False)
 
     def __unicode__(self):
         return str(self.glucose_value)
